Refatora adaptadores em módulos individuais
