+++ sentinela/infrastructure/__init__.py
@@ -6,7 +6,11 @@ Exposes concrete implementations and helpers so consumers can import from
 
 from .database import MongoClientFactory, MongoSettings
 from .extraction import MongoNewsRepository, PostgresExtractionResultWriter
-from .repositories import MongoArticleRepository, MongoPortalRepository
+from .repositories import (
+    MongoArticleReadRepository,
+    MongoArticleRepository,
+    MongoPortalRepository,
+)
 from .scraper import RequestsSoupScraper, Scraper
 
 __all__ = [
@@ -14,6 +18,7 @@ __all__ = [
     "MongoClientFactory",
     "MongoPortalRepository",
     "MongoArticleRepository",
+    "MongoArticleReadRepository",
     "MongoNewsRepository",
     "PostgresExtractionResultWriter",
     "Scraper",
