+++ sentinela/infrastructure/repositories/__init__.py
@@ -1,0 +1,11 @@
+"""Implementações de repositórios baseadas em MongoDB."""
+
+from .mongo_article_read_repository import MongoArticleReadRepository
+from .mongo_article_repository import MongoArticleRepository
+from .mongo_portal_repository import MongoPortalRepository
+
+__all__ = [
+    "MongoPortalRepository",
+    "MongoArticleRepository",
+    "MongoArticleReadRepository",
+]
