+++ sentinela/services/publications/adapters/__init__.py
@@ -1,0 +1,7 @@
+"""Adaptadores HTTP e modelos de entrada do serviço de publicações."""
+
+from .article_batch_payload import ArticleBatchPayload
+from .article_payload import ArticlePayload
+from .ingestion_router import create_ingestion_router
+
+__all__ = ["ArticlePayload", "ArticleBatchPayload", "create_ingestion_router"]
